Module delete validation backend

+++ TutoringApp/Services/Modules/ModulesService.cs
@@ -1,4 +1,6 @@
 ﻿using AutoMapper;
+using Microsoft.Extensions.Logging;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -16,15 +18,18 @@ namespace TutoringApp.Services.Modules
         private readonly IRepository<Module> _modulesRepository;
         private readonly IMapper _mapper;
         private readonly ICurrentUserService _currentUserService;
+        private readonly ILogger<IModulesService> _logger;
 
         public ModulesService(
             IRepository<Module> modulesRepository,
             IMapper mapper,
-            ICurrentUserService currentUserService)
+            ICurrentUserService currentUserService,
+            ILogger<IModulesService> logger)
         {
             _modulesRepository = modulesRepository;
             _mapper = mapper;
             _currentUserService = currentUserService;
+            _logger = logger;
         }
 
         public async Task<IEnumerable<NamedEntityDto>> GetAll()
@@ -50,6 +55,13 @@ namespace TutoringApp.Services.Modules
 
             if (module != null)
             {
+                if (module.ModuleTutors.Any())
+                {
+                    var errorMessage = $"Could not delete module (id='{id}'): it contains some tutors.";
+                    _logger.LogError(errorMessage);
+                    throw new InvalidOperationException(errorMessage);
+                }
+
                 await _modulesRepository.Delete(module);
             }
         }
